requirements.txt
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.3
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.8.2
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.9
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap==0.2.4
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-six==1.0.4
gmane==0.1.dev25
GroupyAPI==0.10.3
gunicorn==20.0.4
httplib2==0.9.2
hyperlink==17.3.1
idna==2.9
incremental==16.10.1
isort==4.3.21
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.0
mccabe==0.6.1
networkx==2.4
numpy==1.18.2
oauthlib==3.1.0
parsedatetime==2.5
Pillow==7.1.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyOpenSSL==17.5.0
pyserial==3.4
python-dateutil==2.8.0
python-debian==0.1.32
python-json-logger==0.1.11
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
pywin32-ctypes==0.2.0
pyxdg==0.25
PyYAML==3.12
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
scipy==1.4.1
SecretStorage==2.3.1
selenium==3.141.0
service-identity==16.0.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
splinter==0.13.0
sqlparse==0.3.0
stripe==2.45.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
whitenoise==5.0.1
wrapt==1.11.2
xlwt==1.3.0
zope.interface==4.3.2
